Track descriptors for all inputs/outputs of AOTAutograd traced graph
One of the recurring challenges of working with FX graphs produced by
AOTAutograd is that there is a very intricate input/output calling
convention that is essentially impossible to understand without actually
reverse engineering the AOTAutograd code.  It is so bad that there
is a bit of logic for stashing indices of relevant arguments/outputs
in TracingContext so Inductor can figure out what the correct arguments
are.

This PR introduces the necessary scaffolding to keep track of
"descriptors" of every input/output to a (joint) FX graph produced
by AOTAutograd.  First read through descriptors.py to get a sense for
what is available: for inputs, you can figure out if you have
a plain input, tangent, parameter, or something more exotic like
one of the fields of a subclass or view base.  For outputs, you can
determine if you have a plain output or grad, or something more exotic
like the contents of a mutated input or an intermediate base of several
views that were returned.

There are two distinct parts of this patch: AOTInput tracking, and
AOTOutput tracking.

**AOTInput tracking.**  The way this works is that AOTAutograd starts of
with some Tensor `flat_args` that are the inputs to the graph being
traced, and then updates these arguments as it modifies the input
calling convention.  Anywhere these `args` are passed around, we now add a
news argument `args_descs` which is updated in synchrony with args.  Add
a new arg?  Add a new AOTInput to `args_descs`.

**AOTOutput tracking.**  Originally, I wanted to also add an `outs_descs`
analogous to `args_descs` tracking output metadata.  However, it is
often difficult to compute what the output will be until you're actually
tracing the function for real (and are able to peek at the real
outputs).  So we only compute `outs_desc` when we actually trace.  To do
this, we change the calling convention of the function we trace to
return not just outputs, but a tuple of `outs` and `outs_descs`.  Before
we bottom out at the `make_fx` invocation, we save `outs_descs` to a
nonlocal and bottom out.

TODO: Obviously, we need to store these results somewhere and then test
them.  My general strategy will be to put it on the meta of the
input/output nodes and do some expect tests for them in the AOTAutograd
tests.  However, this current PR is a good litmus test for whether or
not I've nailed all the plumbing correctly, since I made the changes in
a maximally BC breaking way (new input arguments / new output return
types) which means that if I did anything wrong I should fail quickly.
We potentially should introduce some abstractions so that, for example,
we don't have to keep passing both `args` and `args_descs` around, but
this would have been a more involved refactor, better to do this
incrementally later.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>
ghstack-source-id: 0d4d57f2cee3d0a08ad4f7892470b63eabb79003
Pull-Request: https://github.com/pytorch/pytorch/pull/158624