[OpenReg] Add Develop Notes for Integrating New Backend into PyTorch
As the title stated.

To facilitate the integration of the new backend, we plan to publish a new
development note that details all the key components, hoping to speed up
the development of other accelerators.

This PR is the beginning of this note, which we will gradually improve and
keep in sync with OpenReg's code.

ghstack-source-id: 9ea01fc8e65b1f0f68eeb33714df8b7d2a08f0cc
Pull-Request-resolved: https://github.com/pytorch/pytorch/pull/158644